Implemented basic comment 'delete' feature.
# Board
- 게시글과 댓글 모두 비밀번호가 암호화되어 있기 때문에 이를 검증하는 로직이 필요함.
- 각각 authArticleAccess, authCommentAccess로 명명 및 작성하였음.

## BoardExceptionHandler.java
- 삭제하려는 댓글이 존재하지 않는 예외 정의(NoCommentFoundException)
- 댓글 비밀번호 인증이 실패했을 때 예외 정의(CommentAuthFailedException)

## BasicBoardService.java
- 엔티티의 존재 여부를 확인할 때 기존의 Optional 객체를 받는 findById 메서드 대신 existsById 메서드로 교체.
- 댓글 뿐 아니라 게시글 서비스도 전반적으로 해당 메서드를 적용.

# Command Object
- 댓글 삭제 요청 커맨드 객체 작성.

# Test
## BoardMvcTest.java
- 존재하지 않는 게시글을 삭제하려고 할 때 Validation을 만족하는 경우와 만족하지 않는 경우의 처리가 다르기 때문에 테스트 코드 수정.
- Validation을 만족하는 적절하지 못한 요청은 발생되는 예외에 따라 달리 처리됨.
- Validation을 만족하지 못하는 요청은 현재 개별 메서드에서 BindingResult의 hasErrors 메서드 결과에 따라 대부분 Redirection으로 처리중.
- 댓글 삭제 테스트 코드 작성(비밀번호 오류, 부적절한 댓글 ID 검증).

## CommentControllerTest.java
- 댓글 삭제 테스트 코드 작성(존재하지 않는 댓글 ID 검증).

# TODO
- Validation을 만족하지 못하는 요청은 단순히 Redirection하지 않고 ExceptionHandler를 정의해 둘 것.

+++ src/main/java/com/haruhiism/bbs/service/BasicBoardService.java
@@ -3,6 +3,7 @@ package com.haruhiism.bbs.service;
 import com.haruhiism.bbs.domain.entity.BoardArticle;
 import com.haruhiism.bbs.domain.entity.BoardComment;
 import com.haruhiism.bbs.exception.NoArticleFoundException;
+import com.haruhiism.bbs.exception.NoCommentFoundException;
 import com.haruhiism.bbs.exception.UpdateDeletedArticleException;
 import com.haruhiism.bbs.repository.ArticleRepository;
 import com.haruhiism.bbs.repository.CommentRepository;
@@ -58,10 +59,10 @@ public class BasicBoardService implements ArticleService, CommentService {
     @Override
     @Transactional
     public void updateArticle(BoardArticle boardArticle) {
-        if(articleRepository.findById(boardArticle.getArticleID()).isEmpty()){
-            throw new UpdateDeletedArticleException();
-        } else {
+        if(articleRepository.existsById(boardArticle.getArticleID())){
             articleRepository.save(boardArticle);
+        } else {
+            throw new UpdateDeletedArticleException();
         }
     }
 
@@ -69,7 +70,11 @@ public class BasicBoardService implements ArticleService, CommentService {
     @Override
     @Transactional
     public void deleteArticle(Long articleID){
-        articleRepository.deleteByArticleID(articleID);
+        if(articleRepository.existsById(articleID)) {
+            articleRepository.deleteByArticleID(articleID);
+        } else {
+            throw new NoArticleFoundException();
+        }
     }
 
     @Override
@@ -78,31 +83,41 @@ public class BasicBoardService implements ArticleService, CommentService {
         articleRepository.delete(boardArticle);
     }
 
-
     @Override
-    public boolean authEntityAccess(Long articleID, String rawPassword){
-        BoardArticle readArticle = readArticle(articleID);
-        return dataEncoder.compare(rawPassword, readArticle.getPassword());
+    public boolean authArticleAccess(Long articleID, String rawPassword) {
+        if(articleRepository.existsById(articleID)) {
+            BoardArticle readArticle = readArticle(articleID);
+            return dataEncoder.compare(rawPassword, readArticle.getPassword());
+        } else {
+            throw new NoArticleFoundException();
+        }
     }
+
 
 
     @Override
     @Transactional
     public void createComment(BoardComment comment) {
-//        Optional<BoardArticle> commentedArticle = articleRepository.findById(comment.getArticleID());
-//        if(commentedArticle.isEmpty()){
-//            throw new NoArticleFoundException();
 //        } else {
 //            BoardArticle article = commentedArticle.get();
 //            article.getComments().add(comment);
 //            articleRepository.save(article);
 //        }
-        Optional<BoardArticle> commentedArticle = articleRepository.findById(comment.getArticleID());
-        if(commentedArticle.isEmpty()){
-            throw new NoArticleFoundException();
-        } else {
+        if(articleRepository.existsById(comment.getArticleID())){
             comment.setPassword(dataEncoder.encode(comment.getPassword()));
             commentRepository.save(comment);
+        } else {
+            throw new NoArticleFoundException();
+        }
+    }
+
+    @Override
+    public BoardComment readComment(Long commentID) {
+        Optional<BoardComment> comment = commentRepository.findById(commentID);
+        if(comment.isEmpty()){
+            throw new NoCommentFoundException();
+        } else {
+            return comment.get();
         }
     }
 
@@ -124,19 +139,27 @@ public class BasicBoardService implements ArticleService, CommentService {
 
     @Override
     @Transactional
-    public void deleteAllCommentsOfArticle(Long articleID) {
-
-    }
-
-    @Override
-    @Transactional
     public void deleteComment(Long commentID) {
-
+        if(commentRepository.existsById(commentID)) {
+            commentRepository.deleteById(commentID);
+        } else {
+            throw new NoCommentFoundException();
+        }
     }
 
     @Override
     @Transactional
     public void deleteComment(BoardComment comment) {
+        commentRepository.delete(comment);
+    }
 
+    @Override
+    public boolean authCommentAccess(Long commentID, String rawPassword){
+        if(commentRepository.existsById(commentID)) {
+            BoardComment comment = readComment(commentID);
+            return dataEncoder.compare(rawPassword, comment.getPassword());
+        } else {
+            throw new NoCommentFoundException();
+        }
     }
 }
